requirements.txt
python-docx>=1.1.0
sentence-transformers>=2.7.0
transformers>=4.51.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.3.0
